requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
charset-normalizer>=3.2.0
httpx[http2]>=0.25.0
tenacity>=8.2.0
pydantic>=2.5.0
lxml>=4.9.0
selectolax>=0.3.0
dateparser>=1.2.0
pytz>=2023.3
rapidfuzz>=3.5.0
phonenumbers>=8.13.0
orjson>=3.9.0
sqlalchemy>=2.0.0
structlog>=23.2.0
pdfminer.six>=20221105
pypdf>=3.17.0
openai>=1.0.0
# Advanced PDF generation using Playwright (handles JavaScript, CSS, modern web pages)
playwright>=1.40.0
